+++ my_unreal_dx12/Mesh.h
@@ -14,7 +14,7 @@ struct Vertex {
 
 class Mesh {
 public:
-    Mesh(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices);
+    Mesh(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices);
     Mesh(const std::string& filename);
 
     Mesh(const Mesh&) = default;
